🦄 refactor(travel-on/favorites): Update Server
update logic from server
