+++ apps/travel-on/src/app/favorites/index.ts
@@ -1,2 +1,2 @@
-export { FavoriteHttpService } from './favorite.https.service';
+export { FavoriteHttpService } from './favorite.https.service.copy';
 export { Favorite } from './model';
