+++ apps/travel-on/src/styles.scss
@@ -26,5 +26,5 @@ button {
 }
 
 mat-icon {
-  color: #707070 !important;
+  color: #707070;
 }
